refactor : Car, Result 클래스 리팩토링 (메시지 전달 위주로)
## 목적
 Car 객체가 다른 객체에게 끼치는 의존성을 줄이기 위해 getter를 삭제하고,
 Result 객체를 static 하게 변경하여
 한 라운드가 끝나고 각 자동차에게 메시지를 전달하여 각 라운드별 결과를 출력하기 위함이다.

## 변경
 Car
 - getter의 삭제
 - String whereIsIt() : 해당 자동차가 어디있는지 출력할 문자열을 반환한다.

Result
- field와 생성자 전부 삭제
- static String of(String,int) : name 과 position을 바탕으로 현재 자동차의 위치를 문자열로 만든다.
- static String repeat(int) : position 만큼 '-'를 반복하여 현재 위치를 나타낸다.
