+++ src/main/java/racingcar/Car.java
@@ -13,15 +13,11 @@ public class Car {
         return new Car(name);
     }
 
-    public String getName() {
-        return name;
-    }
-
-    public int getPosition() {
-        return position;
-    }
-
     public void move() {
         position++;
+    }
+
+    public String whereIsIt() {
+        return Result.of(name, position);
     }
 }
